Disconnect kicked user
