+++ app.py
@@ -1,14 +1,13 @@
 from datetime import timedelta
-from typing import Optional
 
 from flask import Flask, make_response, jsonify, request, session
-from flask.sessions import SessionMixin
 from flask_cors import CORS
 from flask_restful import Api
 from flask_socketio import SocketIO
 from jsonschema import ValidationError
 
-from models import db, User
+from helpers.get_user_by_session import get_user_by_session
+from models import db
 from routes.HelloWorld import HelloWorld
 from routes.SessionJoinRoute import SessionJoinRoute
 from routes.SessionPlayerList import SessionPlayerList
@@ -43,26 +42,6 @@ def bad_request(error):
         return make_response(jsonify({'error': original_error.message}), 400)
     # handle other "Bad Request"-errors
     return error
-
-
-def get_user_by_session(ses: SessionMixin) -> Optional[User]:
-    # Get user-id from session
-    user_id = ses.get("user_id")
-
-    # Has session connected
-    if user_id is None:
-        return None
-
-    # Get user from database
-    user: User = User.query.filter_by(id=user_id).first()
-
-    # Check if user exists
-    if user is None:
-        # Session remove 'connected' user
-        session.pop("user_id")
-        return None
-
-    return user
 
 
 @socketio.on('connect')
